tests/requirements.txt
pytest==8.3.5
httpx==0.28.1
requests==2.32.3
Pillow==11.2.1
reportlab==4.4.0
python-docx==1.1.2
openpyxl==3.1.5
python-pptx==1.0.2
